chore: board and food

+++ src/board.rs
@@ -9,13 +9,13 @@ pub const TILE_SIZE: f32 = 30.0;
 #[derive(Resource, Debug)]
 pub struct Board {
     // the board will be a square, (e.g. if size is 5 then the board is 5^2 or 25 tiles)
-    pub size: usize,
+    pub size: i32,
     // @info: size * tile size -> the pixel quantity to be used when rendering the board
     pub physical_size: f32,
 }
 
 impl Board {
-    fn new(size: usize) -> Self {
+    fn new(size: i32) -> Self {
         let physical_size: f32 = size as f32 * TILE_SIZE;
         Self {
             size,
@@ -26,7 +26,7 @@ impl Board {
     // @todo: implement function to turn a board's cell position into the physical rendered board.
     // Since the original board sizing uses an int 20 for example, we need to multiply that with
     // the tile size to get the physical size of the board to be rendered in pixels
-    pub fn position_translate(&self, pos: usize) -> f32 {
+    pub fn position_translate(&self, pos: i32) -> f32 {
         let offset = -&self.physical_size / 2.0 + 0.5 * TILE_SIZE;
         offset + pos as f32 * TILE_SIZE
     }
